aio: Silence integer conversion warning
Commit "aio: Change aio_context_t to a user pointer" ensured that
the compat io_setup handler always manipulates the context as
compat_aio_context_t. However, that type is 32-bit in compat32 and
printing it with %lu will trigger a warning. Add a cast to unsigned
long to avoid that warning.

+++ fs/aio.c
@@ -1470,7 +1470,7 @@ COMPAT_SYSCALL_DEFINE2(io_setup, unsigned, nr_events, compat_aio_context_t __use
 	ret = -EINVAL;
 	if (unlikely(ctx || nr_events == 0)) {
 		pr_debug("EINVAL: ctx %lu nr_events %u\n",
-		         ctx, nr_events);
+		         (unsigned long)ctx, nr_events);
 		goto out;
 	}
